fix: use `rememberUpdatedState` on all state used in `pointerInput`
Either this or supplying them as keys to `pointerInput` is necessary (which would recreate the entire set of lambdas), otherwise an updated parameter will not get updated in the lambda (in this case, the callbacks). This would make using state from within the lambda impossible unless the user manually wraps the state used in callback lambda with rememberUpdatedState`, which isn't documented anywhere and seems like buggy behavior.